backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt==3.1.7
geoalchemy2>=0.14.0
alembic>=1.11.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
httpx>=0.24.0
redis>=5.0.0
pandas>=2.0.0
openpyxl>=3.1.0
loguru>=0.7.0
asyncpg>=0.28.0
email-validator>=2.0.0
slowapi>=0.1.9
firebase-admin>=6.2.0
supabase>=2.4.0
sentry-sdk[fastapi]>=1.14.0
